+++ Sources/Views/ALKChatBar.swift
@@ -105,8 +105,8 @@ open class ALKChatBar: UIView, Localizable {
 
     open lazy var placeHolder: UITextView = {
         let view = UITextView()
-        view.setFont(UIFont.font(.normal(size: 14)))
-        view.setTextColor(.color(Color.Text.gray9B))
+        view.setFont(ALKChatBarConfiguration.TextView.placeholder.font)
+        view.setTextColor(ALKChatBarConfiguration.TextView.placeholder.text)
         view.text = localizedString(forKey: "ChatHere", withDefaultValue: SystemMessage.Information.ChatHere, fileName: configuration.localizedStringFileName)
         view.isUserInteractionEnabled = false
         view.isScrollEnabled = false
@@ -224,12 +224,13 @@ open class ALKChatBar: UIView, Localizable {
         }
     }
 
-    var defaultTextAttributes: [NSAttributedString.Key: Any] = {
+    lazy var defaultTextAttributes: [NSAttributedString.Key: Any] = {
         let style = NSMutableParagraphStyle()
         style.lineSpacing = 4.0
         let attrs = [
             NSAttributedString.Key.paragraphStyle: style,
-            NSAttributedString.Key.font: UIFont.font(.normal(size: 16.0)),
+            NSAttributedString.Key.font: ALKChatBarConfiguration.TextView.text.font,
+            NSAttributedString.Key.foregroundColor: ALKChatBarConfiguration.TextView.text.text,
         ]
         return attrs
     }() {